requirements.txt
discord.py==2.4.0
wavelink==3.4.1
spotipy==2.25.1
aiohttp==3.9.5
python-dotenv==1.0.1
gTTS==2.5.1
PyNaCl==1.5.0
